// src/app/models/enamel.ts
export class Enamel{
    enamelId: number = 0;
    enamelSku: string = '';
    enamelName: string = '';
    pantoneColor: string = '';
    hardness: boolean = false;
    supplierId: number = 0;
    colorHex: string = '';
}

export class EnamelColumns{
    'ID': number = 0;
    'SKU': string = '';
    'Name': string = '';
    'Pantone Color': string = '';
    'Hardness': boolean = false;
    'Supplier': string = '';
    'Color': string = '';
}